feat: change the worker metadata to be dependent of the worker
At the moment, I think that the workflow steps should have access
to the engine internal such as the used MongoDB connection or
MongoDB database.
This way, it would simplify the end-user life as he won't have
to connect again on the database.
At the end, I want to also share the db transaction to commit
eventual changes in db made by the step and the workflow state change.