feat: support workflow made of multiple steps
The code is a bit messy for the moment, especially the communication
with mongodb. I'll implement the same design to handler storage/polling
as I did with the previous implementation, in order to tackle this
complexity.
I'll probably start by implementing an abstraction for the storage,
then an abstraction for the poller.