add a onStepCompleted hook
I've skipped the test because I need to refactor
how hooks are executed.
For the moment, only the current workflow context's hooks
are executed. But, what if a hook was defined in a plugin.

To execute hooks correctly I need to do something similar
to the what I've done with the step execution graph:
Register the hook on each context, climb the prototype chain
to get the root workflow, and collect the hooks in the each
child context, but iterating depth-first.