make the plugin dependencies verification work
For the moment the plugin dependencies' verification works,
but it is not ideal.
First, the iteration from bottom to top could be optimized (I think using a while loop).
Second, the error does not tell which plugin has an issue and which of its dependency is missing.

Also, the `register` method has a issue: if I implement async plugin,
then the `register` method will be colored, making the API less user friendly.