create a alternative pull model
I want to be able to display the workflows in a UI in order
to track their progress but also to perform actions, such as retrying.

Its the reason why I need a better model that include information about
the task to perform (under the `toExecute` property), but also general
data about the workflow such as the steps' names, the attempts and the
workflow's status.