make test pass except for the retry tests
The issue is that I cannot use the filter pickAfter <= NOW
in the watch stream pipeline. So, I'll have to start a poller
to handle the failed messages. Additionaly I'll have to start
another poller that handles the messages missed by the worker
while it was not running (e.g the worker has crashed, but the
workflows were still trigered).