implement another style of consumption
I tried to implement a RabbitMQ engine but the way RabbitMQ ack
works, makes it hard to work with an async iterator. So, I'm using
a callback function that is executed by the consumption.

Also, I've created a compound consumption to hide the fact that
multiple consumption are used at the same time.