remove the resilience test
The test suite is too flacky because of amqplib.
I'll implement the reconnect logic without tests.
Once I'll have nothing else to do, I'll try to fix the issue
in the amqplib package.

For the record, I have from time to time, pending promise that are
not resolved before the test is ended.
I'm pretty sure that is the fault of amqplib.